Windows: Use memmove instead of bcopy
The bcopy is deprecated (marked as LEGACY in
POSIX.1-2001): use memcpy(3) or memmove(3) in new programs.  Note  that
the  first two arguments are interchanged for memcpy(3) and memmove(3).
POSIX.1-2008 removes the specification of bcopy().

The  memmove()  function  copies n bytes from memory area src to memory
area dest.  The memory areas may overlap: copying takes place as though
the  bytes in src are first copied into a temporary array that does not
overlap src or dest, and the bytes are then copied from  the  temporary
array to dest.